requirements.txt
flask
numpy
p pillow
onnxruntime
gspread
oauth2client
python-dotenv
requests
pytz
